refactor: layered architecture
