+++ src/application/entity.ts
