+++ src/application/presentation/serialization/messageSerializer.ts
