+++ src/application/room/user.ts
