+++ src/presentation/commands/userToken.ts
@@ -1,8 +1,8 @@
 import { Server, Socket } from 'socket.io'
 import { commandListener } from '../utils'
 import { joinCommand } from './joinRoom'
-import { Ack } from '../../model/message'
-import { RoomAggregate } from '../../aggregates/room/roomAggregate'
+import { Ack } from '../../application/message'
+import { RoomService } from '../../application/roomService'
 import { Commands } from './commands'
 
 /**
@@ -17,7 +17,7 @@ import { Commands } from './commands'
 export function userTokenCommand(
   io: Server,
   socket: Socket,
-  roomController: RoomAggregate
+  roomController: RoomService
 ): (message: any, ack: any) => void {
   return (message, ack) => {
     const { token } = message
