+++ src/presentation/defineNamespaces.ts
@@ -1,7 +1,7 @@
 import { Server } from 'socket.io'
 import { connectionCommand } from './commands/connection'
-import { RoomAggregate } from '../aggregates/room/roomAggregate'
+import { RoomService } from '../application/roomService'
 
 export async function registerCommands(io: Server) {
-  connectionCommand(io, new RoomAggregate())
+  connectionCommand(io, new RoomService())
 }
